requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
clearbit==0.1.7
coreapi==2.3.3
coreschema==0.0.4
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drfdocs==0.0.11
email-hunter-python==1.1.0
get==2018.11.19
itypes==1.1.0
Jinja2==2.10
kombu==4.2.2.post1
MarkupSafe==1.1.0
openapi-codec==1.3.2
pipreqs==0.4.9
post==2018.11.20
public==2018.11.20
PyJWT==1.7.1
python-dotenv==0.10.1
query-string==2018.11.20
redis==3.1.0
request==2018.11.20
simplejson==3.16.0
uritemplate==3.0.0
vine==1.2.0
yarg==0.1.9
